CartItemRepository 리팩토링

+++ src/main/java/kwh/cofshop/cart/service/CartItemService.java
@@ -10,6 +10,7 @@ import kwh.cofshop.cart.repository.CartItemRepository;
 import kwh.cofshop.cart.repository.CartRepository;
 import kwh.cofshop.global.exception.BusinessException;
 import kwh.cofshop.global.exception.errorcodes.BusinessErrorCode;
+import kwh.cofshop.global.exception.errorcodes.UnauthorizedErrorCode;
 import kwh.cofshop.item.domain.Item;
 import kwh.cofshop.item.domain.ItemOption;
 import kwh.cofshop.item.repository.ItemOptionRepository;
@@ -45,6 +46,7 @@ public class CartItemService {
 
         // 중복을 제거하고 수량을 합산하기 위해서 Map 선언
         Map<String, CartItemRequestDto> mergedCartItems = new HashMap<>();
+
 
         for (CartItemRequestDto requestDto : cartItemRequestDto) {
             String key = requestDto.getItemId() + "-" + requestDto.getOptionId();
@@ -92,13 +94,24 @@ public class CartItemService {
         return responseDto;
     }
 
-    // 장바구니 아이템 삭제
     @Transactional
-    public void deleteCartItem(Long cartItemId) {
-        CartItem cartItem = cartItemRepository.findById(cartItemId)
+    public void deleteCartItem(Long memberId, Long cartId, Long itemId, Long optionId) {
+        // 현재 로그인한 사용자의 장바구니인지 검증
+        Cart cart = cartRepository.findById(cartId)
+                .orElseThrow(() -> new BusinessException(BusinessErrorCode.CART_NOT_FOUND));
+
+        // 만약 로그인 된 사용자가 다르다면 오류 -> 로그아웃하고 다른 아이디로 로그인 해서 삭제한다면? 삭제당할 위험 존재
+        if (!cart.getMember().getId().equals(memberId)) {
+            throw new BusinessException(UnauthorizedErrorCode.MEMBER_UNAUTHORIZED);
+        }
+
+        CartItem cartItem = cartItemRepository.findByItemOptionIdAndItemIdAndCartId(
+                        cartId, itemId, optionId)
                 .orElseThrow(() -> new BusinessException(BusinessErrorCode.ITEM_NOT_FOUND));
+
         cartItemRepository.delete(cartItem);
     }
+
     // 장바구니 전체 삭제
     @Transactional
     public void deleteCartItemAll(Long cartId) {
